refactor: PerformanceSeatInfoRespository id값 Integer로 수정

+++ src/main/java/com/wanted/preonboarding/ticket/infrastructure/repository/PerformanceSeatInfoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.Optional;
 import java.util.UUID;
 
-public interface PerformanceSeatInfoRepository extends JpaRepository<PerformanceSeatInfo, Long> {
+public interface PerformanceSeatInfoRepository extends JpaRepository<PerformanceSeatInfo, Integer> {
     @Query("select ps from PerformanceSeatInfo ps where ps.performanceId =:performanceId and ps.isReserve =:isReserve")
     Optional<PerformanceSeatInfo> findByPerformanceIdAndIsReserve(UUID performanceId, String isReserve);
 }
